requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
PyYAML==6.0
requests==2.28.2
soupsieve==2.4.1
tqdm==4.65.0
urllib3==1.26.15
yarl==1.9.1